Commit All

+++ cop WebApi/Controllers/ChildController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Cors;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.VisualBasic;
 using OfficeOpenXml;
 using System.ComponentModel;
 
@@ -67,6 +68,9 @@ namespace Co_p_new__WebApi.Controllers
                         var parent1 = worksheet.Cells[row, 6].Text;
                         var parent2 = worksheet.Cells[row, 7].Text;
                         var childPhotoName = worksheet.Cells[row, 8].Text;
+                        var CurrentAcademicYear = int.Parse(worksheet.Cells[row, 9].Text);
+                        var KindergartenNumber = int.Parse(worksheet.Cells[row, 10].Text);
+
 
                         // Retrieve or create related entities as needed
                         var child = db.Children.FirstOrDefault(c => c.ChildId == childId);
@@ -86,6 +90,7 @@ namespace Co_p_new__WebApi.Controllers
 
                             children.Add(newChild);
                         }
+
                     }
                 }
             }
@@ -119,7 +124,11 @@ namespace Co_p_new__WebApi.Controllers
         [Route("DeleteChild")]
         public dynamic DeleteChild(string ID)
         {
-            Child c = db.Children.Where(x => x.ChildId == ID).FirstOrDefault();
+            Child? c = db.Children.Where(x => x.ChildId == ID).FirstOrDefault();
+            if (c == null)
+            {
+                return ("Child not found");
+            }
             db.Children.Remove(c);
             db.SaveChanges();
             return (c.ChildFirstName + " " + c.ChildSurname + "deleted");
